backend/requirements.txt
crewai
crewai-tools
langchain-openai
fastapi
uvicorn[standard]
python-dotenv
duckduckgo-search